requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
djangorestframework
django-filter
